+++ changelog.md
@@ -22,6 +22,9 @@ TODO: Not debugged:
 - Death
 - Transition
 
+## 6 March 2025
+- Futher minor improvements to `removeFromFlexleSampler!` dict update (CLM)
+
 ## 5 March 2025
 - Add response acquisition upon clearance as preferred alternative to response
 acquisition during infection (as in mutations upon infection vs. mutation